Implement the behavior for converting to lexicoids
Lexicoids are special cases of Base32-encoding for integers. These
integers correspond to timestamps, which naturally come from the Unix
epoch.

This change handles timestamps of many forms and converts them into
lexicoids appropriately.

The test suite conforms to the same list of tests from [the Rust
library][1] in addition to a few edge cases and a verification of one
from Brandur's blog.

[1]: https://github.com/lmammino/lexicoid/blob/9efe314998609a769aff405d16a0e4e95dab13de/src/lib.rs#L100-L111