Update the checksum for 0.1.0
This was a :facepalm: on my part. I discovered this feature and started
running `rake build:checksum` and checking that file into version
control. However, I did not realize that when you then run `rake
release`, it _rebuilds_ the gem so the checksum won't match anymore.

This is the actual checksum as the gem that I built and pushed to
RubyGems.org.